refactor: Images layout to use react-query
